refactor: modularize modal, create 2 components for login/signup
