+++ src/client/components/index.ts
@@ -6,3 +6,5 @@ export { default as Modal } from './Modal/Modal';
 export { default as Recipe } from './Recipe/Recipe';
 export { default as RecipeContainer } from '../containers/RecipeContainer/RecipeContainer';
 export { default as DropDown } from './DropDown/DropDown';
+export { default as LoginForm } from './LoginForm/LoginForm';
+export { default as SignupForm } from './SignupForm/SignupForm';
